Increase available disk space for data
Before this change the pod could only access the boot disk for
storage, and this ended up being too small for some workloads.
Instead of increasing the boot disk size, this change adds
ephemeral storage in the form of an emptyDir volume, which should
prove to be more flexible in the long run. The extra storage is
discarded when a Pod leaves a Node for any reason.
I decided to add ephemeral storage instead of persistent storage
since this avoids having to make sure that all sensitive data is
deleted. This storage is taken from the local disk of the node.
To keep things simple this emptyDir will contain all of the data
and nothing else. Everything else will still be copied
and installed etc. to the home dir of the Docker container.